requirements.txt
# Core
huggingface-hub>=0.23.0
requests>=2.31.0
pyyaml>=6.0

# Notebooks & EDA
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0

# Optional: Streamlit app
streamlit>=1.32.0

# Dev / testing
pytest>=8.0.0
